Rename libbess-python to pybess
Python packages should not have "-" in their name.
