+++ bessctl/bessctl
@@ -15,10 +15,10 @@ import commands
 
 try:
     this_dir = os.path.dirname(os.path.realpath(__file__))
-    sys.path.insert(1, '%s/../libbess-python' % this_dir)
-    from bess import *
+    sys.path.insert(1, os.path.join(this_dir, '..'))
+    from pybess.bess import *
 except ImportError:
-    print >> sys.stderr, 'Cannot import the API module (libbess-python)'
+    print >> sys.stderr, 'Cannot import the API module (pybess)'
     raise
